cards

+++ Api/LancacheManager/Models/CacheInfo.cs
@@ -20,6 +20,6 @@ public class CacheInfo
     public long UsedCacheSize { get; set; }
     public long FreeCacheSize { get; set; }
     public double UsagePercent => TotalCacheSize > 0 ? (UsedCacheSize * 100.0) / TotalCacheSize : 0;
-    public int TotalFiles { get; set; }
+    public long TotalFiles { get; set; }
     public Dictionary<string, long> ServiceSizes { get; set; } = new();
 }
